+++ pubspec.yaml
@@ -40,6 +40,7 @@ dependencies:
   persistent_bottom_nav_bar_v2: ^5.3.1
   device_preview: ^1.2.0
   google_maps_flutter: ^2.9.0
+  flutter_map: ^7.0.2
 
 dev_dependencies:
   flutter_test:
